+++ infrastructures/doctrine/DBSource/ODM/ExprConversionTrait.php
@@ -25,9 +25,11 @@ declare(strict_types=1);
 
 namespace Teknoo\East\Website\Doctrine\DBSource\ODM;
 
+use Teknoo\East\Website\Doctrine\DBSource\Common\ExprConversionTrait as CommonTrait;
 use Teknoo\East\Website\Query\Expr\ExprInterface;
 use Teknoo\East\Website\Query\Expr\In;
 use Teknoo\East\Website\Query\Expr\InclusiveOr;
+use Teknoo\East\Website\Query\Expr\ObjectReference;
 
 /**
  * @license     http://teknoo.software/license/mit         MIT License
@@ -35,30 +37,30 @@ use Teknoo\East\Website\Query\Expr\InclusiveOr;
  */
 trait ExprConversionTrait
 {
-    private function convert(array &$values): array
+    use CommonTrait;
+
+    /**
+     * @param In $expr
+     */
+    private static function processInExpr(array &$final, string $key, ExprInterface $expr): void
     {
-        $final = [];
-        foreach ($values as $key => $value) {
-            if (!$value instanceof ExprInterface) {
-                $final[$key] = $value;
+        $final[$key] = ['$in' => $expr->getValues()];
+    }
 
-                continue;
-            }
+    /**
+     * @param InclusiveOr $expr
+     */
+    private static function processInclusiveOrExpr(array &$final, string $key, ExprInterface $expr): void
+    {
+        $orValues = $expr->getValues();
+        $final['$or'] = static::convert($orValues);
+    }
 
-            if ($value instanceof In) {
-                $final[$key] = ['$in' => $value->getValues()];
-
-                continue;
-            }
-
-            if ($value instanceof InclusiveOr) {
-                $orValues = $value->getValues();
-                $final['$or'] = $this->convert($orValues);
-
-                continue;
-            }
-        }
-
-        return $final;
+    /**
+     * @param ObjectReference $expr
+     */
+    private static function processObjectReferenceExpr(array &$final, string $key, ExprInterface $expr): void
+    {
+        $final[$key . '.$id'] = $expr->getObject()->getId();
     }
 }